Enable variadic macros for PGI compilers
Remove PGI compilers with an EDG front end from the list of untested
compilers for which BOOST_PP_VARIADICS is always false.  The PGI C++
compiler implements variadic macros correctly in C++11 mode or later.
(BOOST_PP_VARIADICS is still false for the PGI C compiler, which is not
EDG based and which has an incomplete implementation of variadic
macros.)